// src/app/core/interfaces/wish-list.ts
export interface WishList {
    status: string
    message: string
    data: string[]
  }
  
  // all 
  export interface AllWishListData {
    status: string
    count: number
    data: AllWishList[]
  }
  
  export interface AllWishList {
    sold: number
    images: string[]
    subcategory: Subcategory[]
    ratingsQuantity: number
    _id: string
    title: string
    slug: string
    description: string
    quantity: number
    price: number
    imageCover: string
    category: Category
    brand: Brand
    ratingsAverage: number
    createdAt: string
    updatedAt: string
    __v: number
    id: string
    priceAfterDiscount?: number
  }
  
  export interface Subcategory {
    _id: string
    name: string
    slug: string
    category: string
  }
  
  export interface Category {
    _id: string
    name: string
    slug: string
    image: string
  }
  
  export interface Brand {
    _id: string
    name: string
    slug: string
    image: string
  }
  